% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmapr-package.R
\docType{package}
\name{nmapr}
\alias{nmapr}
\alias{nmapr-package}
\title{Perform Network Discovery and Security Auditing with 'nmap'}
\description{
Perform Network Discovery and Security Auditing with 'nmap'
}
\author{
Bob Rudis (bob@rud.is)
}
