% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-scripts.R
\name{read_script_description}
\alias{read_script_description}
\alias{nse_descr}
\title{Retrieve descriptioin for an NSE script}
\usage{
read_script_description(nse_file, cat = TRUE)

nse_descr(nse_file, cat = TRUE)
}
\arguments{
\item{nse_file}{name (with or without extension) of the NSE script}

\item{cat}{if \code{FALSE} the NSE script description will not be displayed to the console}
}
\description{
Retrieve descriptioin for an NSE script
}
